# shared/contracts/openapi.nexus-core.snapshot.yaml
openapi: 3.0.3
info:
  title: Nexus Core API
  version: "1.1.0"
  description: |
    Multi-tenant Agent Tool Gateway for controlled tool execution.

    Authentication supports API key (`X-NEXUS-GATEWAY-KEY`) and optional JWT bearer mode (`Authorization: Bearer`) when enabled by feature flags.

    Idempotency semantics for `POST /v1/run` on WRITE tools:
    - Same `Idempotency-Key` + same request fingerprint + `COMPLETED` => replay cached success.
    - Same `Idempotency-Key` + same fingerprint + `IN_PROGRESS` => `IDEMPOTENCY_IN_PROGRESS`.
    - Same `Idempotency-Key` + different fingerprint => `IDEMPOTENCY_KEY_CONFLICT`.
    - Same `Idempotency-Key` + same fingerprint + `FAILED` => terminal replay of cached error.
    - To retry after `FAILED`, clients MUST use a new `Idempotency-Key`.
servers:
  - url: http://localhost:8080
...
paths:
...
  /v1/audit:
    get:
      tags: [Audit]
      operationId: queryAudit
      summary: Query audit events
      security:
        - NexusApiKey: []
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AuditToolName"
        - $ref: "#/components/parameters/AuditDecision"
        - $ref: "#/components/parameters/AuditStatus"
        - $ref: "#/components/parameters/AuditFrom"
        - $ref: "#/components/parameters/AuditTo"
        - $ref: "#/components/parameters/AuditLimit"
      responses:
        "200":
          description: Audit items.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditListResponse"
        "400":
          description: Invalid filters.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/audit/export:
    get:
      tags: [Audit]
      operationId: exportAudit
      summary: Stream audit export for SIEM ingestion
      description: |
        Streams events in ascending `created_at` order.

        - `format=jsonl`: `application/x-ndjson` stream, one JSON object per line.
        - `format=csv`: `text/csv` stream with a fixed header.
      security:
        - NexusApiKey: []
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AuditExportFormat"
        - $ref: "#/components/parameters/AuditToolName"
        - $ref: "#/components/parameters/AuditDecision"
        - $ref: "#/components/parameters/AuditStatus"
        - $ref: "#/components/parameters/AuditFrom"
        - $ref: "#/components/parameters/AuditTo"
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 5000
            default: 1000
      responses:
        "200":
          description: Streamed export payload.
          content:
            application/x-ndjson:
              schema:
                type: string
              examples:
                jsonl:
                  value: |
                    {"request_id":"req_1","tool_name":"transfer","decision":"deny","status":"blocked","event_hash":"hash_1","hash_algo":"sha256"}
                    {"request_id":"req_2","tool_name":"transfer","decision":"allow","status":"success","event_hash":"hash_2","prev_event_hash":"hash_1","hash_algo":"sha256"}
            text/csv:
              schema:
                type: string
              examples:
                csv:
                  value: |
                    created_at,org_id,request_id,tool_name,actor,role,scopes,decision,status,latency_ms,error_code,idempotency_present,idempotency_outcome,timeout_ms,budget_remaining_ms_at_execute,stage_durations_ms,dlp_summary,input,context,output,prev_event_hash,event_hash,hash_algo
                    2026-01-10T08:00:00Z,11111111-1111-1111-1111-111111111111,req_1,transfer,agent-1,bot,"[""tools:run""]",deny,blocked,7,POLICY_DENIED,true,NEW,10000,8800,"{""schema_validation"":1}","{""credit_card"":{""count"":1}}","{""amount"":500}","{""user_id"":""u1""}",null,,hash_1,sha256
        "400":
          description: Invalid format or time range.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/tools/{name}/egress-rules:
    get:
      tags: [Egress]
      operationId: listToolEgressRules
      summary: List host allowlist rules
      security:
        - NexusApiKey: []
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ToolNamePath"
      responses:
        "200":
          description: Egress rules list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/EgressRule"
                required: [items]
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags: [Egress]
      operationId: upsertToolEgressRule
      summary: Upsert host allowlist rule
      description: |
        Default egress policy is deny-all. Tools MUST have at least one egress rule
        to make outbound calls. Only explicitly allowed hosts are reachable.
      security:
        - NexusApiKey: []
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ToolNamePath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EgressRule"
      responses:
        "204":
          description: Upserted.
        "400":
          description: Validation error (empty host).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [Egress]
      operationId: deleteToolEgressRule
      summary: Delete egress rule by host
      security:
        - NexusApiKey: []
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ToolNamePath"
        - name: host
          in: query
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted.
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/assistant/query:
    post:
      tags: [A2A]
      operationId: queryAssistant
      summary: Query operator assistant through nexus-core proxy
      security:
        - NexusApiKey: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantQueryRequest"
      responses:
        "200":
          description: Structured assistant response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantQueryResponse"
components:
  securitySchemes:
    NexusApiKey:
      type: apiKey
      in: header
      name: X-NEXUS-GATEWAY-KEY
      description: Tenant API key (stored hashed in DB).
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Optional JWT+JWKS auth mode when enabled by feature flags.
  parameters:
    ToolNamePath:
      name: name
      in: path
      required: true
      schema:
        type: string
...
    AuditToolName:
      name: tool_name
      in: query
      schema:
        type: string
    AuditDecision:
      name: decision
      in: query
      schema:
        type: string
        enum: [allow, deny]
    AuditStatus:
      name: status
      in: query
      schema:
        type: string
        enum: [success, error, blocked]
    AuditFrom:
      name: from
      in: query
      schema:
        type: string
        format: date-time
    AuditTo:
      name: to
      in: query
      schema:
        type: string
        format: date-time
    AuditLimit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    AuditExportFormat:
      name: format
      in: query
      schema:
        type: string
        enum: [jsonl, csv]
        default: jsonl
  schemas:
    ErrorObject:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required: [code, message]
    ErrorResponse:
      type: object
      properties:
        request_id:
          type: string
          format: uuid
        error:
          $ref: "#/components/schemas/ErrorObject"
      required: [request_id, error]
...
    AuditEvent:
      type: object
      properties:
        request_id:
          type: string
        org_id:
          type: string
          format: uuid
        tool_name:
          type: string
        actor:
          type: string
          nullable: true
        role:
          type: string
          nullable: true
        scopes:
          type: array
          items:
            type: string
        decision:
          type: string
          enum: [allow, deny]
        status:
          type: string
          enum: [success, error, blocked]
        reason:
          type: string
          nullable: true
        latency_ms:
          type: integer
        idempotency_present:
          type: boolean
        idempotency_outcome:
          type: string
        timeout_ms:
          type: integer
          nullable: true
        budget_remaining_ms_at_execute:
          type: integer
          nullable: true
        stage_durations_ms:
          type: object
          additionalProperties:
            type: integer
            format: int64
        created_at:
          type: string
          format: date-time
        input:
          type: object
          additionalProperties: true
          nullable: true
        context:
          type: object
          additionalProperties: true
          nullable: true
        output:
          type: object
          additionalProperties: true
          nullable: true
        dlp_summary:
          type: object
          additionalProperties: true
          nullable: true
        error:
          $ref: "#/components/schemas/ErrorObject"
        prev_event_hash:
          type: string
          nullable: true
        event_hash:
          type: string
          nullable: true
        hash_algo:
          type: string
      required: [request_id, tool_name, decision, status, latency_ms, created_at]
    AuditListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AuditEvent"
      required: [items]
...
    EgressRule:
      type: object
      properties:
        host:
          type: string
        enabled:
          type: boolean
      required: [host]
...
    AssistantQueryRequest:
      type: object
      properties:
        query:
          type: string
      required: [query]
    AssistantQueryResponse:
      type: object
      properties:
        summary:
          type: string
        tables:
          type: array
          items:
            type: object
            properties:
              title:
                type: string
              columns:
                type: array
                items:
                  type: string
              rows:
                type: array
                items:
                  type: object
                  additionalProperties:
                    type: string
        actions:
          type: array
          items:
            type: object
            properties:
              label:
                type: string
              action_type:
                type: string
              payload:
                type: object
                additionalProperties: true
      required: [summary]